+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Next Release
 * Allow empty sequence expressions `seq()`, `pseq()` (#159)
 * Add `no_wrap` option to `head()`, `head_option()`, `first()`, `last()` and `last_option()`, as well as to `seq()`, `pseq()` and `Sequence` constructor
+* Add the `split` transformation
 
 ## Release 1.3.0
 * added precompute attribute to reverse transformation (#137)